fix errors

+++ database/migrations/2023_08_29_125400_create_order_details_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->integer('quantity');
             $table->decimal('totalPrice',10,2);
             // copone_id is use (it isn't require for taht i don't foregin key)
-            $table->bigIncrements('copone_id');
+            $table->integer('copone_id');
             $table->timestamps();
         });
     }
